refactor: componentize input area
